user auth and routes fixed

+++ client/src/BBC.js
@@ -50,10 +50,13 @@ import {withButton} from './ButtonProvider'
    
 
     handleErase = () => {
-            
+        var answer = window.confirm("This will permanently delete you account, are you sure you want to proceed?")
+    if(answer){
         this.props.handleDelete2(this.props.user._id)
         this.props.logout()
         this.props.deleteAll(this.props.user._id)
+    }
+    
     }
